requirements.txt
click>=8.1.0
mutagen>=1.47.0
librosa>=0.10.1
pydub>=0.25.1
numpy>=1.24.0
scipy>=1.11.0
soundfile>=0.12.1
colorama>=0.4.6
rich>=13.0.0
PyYAML>=6.0